requirements.txt
streamlit==1.25.0
Pillow==9.0.1
